Move registration to commands package
