+++ mcmd/commands/add.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import mcmd.config.config as config
 from mcmd import io
-from mcmd.arguments import arguments
+from mcmd.commands._registry import arguments
 from mcmd.client.molgenis_client import post, get, post_files
 from mcmd.command import command
 from mcmd.io import highlight
