+++ mcmd/commands/delete.py
@@ -3,7 +3,7 @@ Deletes an entityType or data from an entityType.
 """
 import mcmd.config.config as config
 from mcmd import io
-from mcmd.arguments import arguments
+from mcmd.commands._registry import arguments
 from mcmd.client.molgenis_client import delete, delete_data, ensure_resource_exists, ResourceType
 from mcmd.command import command
 from mcmd.io import highlight
