+++ mcmd/commands/disable.py
@@ -1,6 +1,6 @@
 import mcmd.config.config as config
 from mcmd import io
-from mcmd.arguments import arguments
+from mcmd.commands._registry import arguments
 from mcmd.client.molgenis_client import ResourceType, post, ensure_resource_exists
 from mcmd.command import command
 from mcmd.io import highlight
