+++ mcmd/commands/give.py
@@ -5,7 +5,7 @@ principal doesn't exist, the program will terminate.
 """
 
 from mcmd import io
-from mcmd.arguments import arguments
+from mcmd.commands._registry import arguments
 from mcmd.client.molgenis_client import grant, PrincipalType, principal_exists, \
     resource_exists, ResourceType, ensure_resource_exists, ensure_principal_exists
 from mcmd.command import command
