+++ mcmd/commands/history.py
@@ -1,6 +1,6 @@
 from mcmd import history as hist
 from mcmd import io
-from mcmd.arguments import arguments
+from mcmd.commands._registry import arguments
 from mcmd.command import command
 from mcmd.logging import get_logger
