+++ mcmd/commands/make.py
@@ -1,6 +1,6 @@
 import mcmd.config.config as config
 from mcmd import io
-from mcmd.arguments import arguments
+from mcmd.commands._registry import arguments
 from mcmd.client.molgenis_client import post, get
 from mcmd.command import command
 from mcmd.io import highlight
